Make movement "FPS-independent", re-enable VSync toggle
Add an FPS_CONSTANT that is deltatime multiplied by 60, which is the FPS we usually
target, and then multiply all movement updates by this constant. Also re-enable the
VSync toggle, as movements work fine without VSync now.

TODO: There are some rounding errors which cause the player to "stick" to objects, as
part of the player box will clip through them. Only seems to occur when sharing screens
on Skype.